Fix up the broken renderer code
The reason why the renderer approach didn't work is because the renderer
API only took one parameter, which was an array of floats. However, the
array of floats was not of a fixed size. Behind the scenes, a "float[]"
in C++ and C is the same thing as a "float*". Therefore, when you do
"sizeof(float*)" in code, you end up with the size being 8 (because
pointers are 8 bytes in x64).

The renderer needs to be told how big the array is. This PR contains
code that changes the renderer API so that the size of the array is also
passed in. This means that open GL is told the correct size of the
vertex array, and hence renders things correctly.

I also removed the 'struct' from the renderer, and added a few extra
pragmas to the header files. I also removed a bunch of unnecessary casts
from the code. We can discuss the const issue later when we have time.